dom/workers/test/serviceworkers/test_get_serviced_advanced.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html>
<head>
  <title>Bug 982726 - test get_serviced </title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none"></div>
<pre id="test"></pre>
<script class="testbody" type="text/javascript">
  var opened = [];
  var registration;

  function start() {
    return navigator.serviceWorker.register("get_serviced_worker_advanced.js",
                                            { scope: "./sw_clients/" })
      .then((swr) => registration = swr);
  }

  function closeAndUnregister() {
    var content = document.getElementById("content");
    ok(content, "Parent exists.");

    while (opened.length) {
      content.removeChild(opened.pop());
    }

    return registration.unregister().then(function(result) {
      ok(result, "Unregister should return true.");
    });
  }

  function testGetServiced() {
    var p = new Promise(function(res, rej) {
      window.onmessage = function(e) {
        if (e.data === "READY") {
          if (!registration.active) {
            ok(false, "Did not have active worker");
            return Promise.resolve();
          }

          opened.push(iframe);
          registration.active.postMessage("getServiced");
        } else if (e.data === opened.length) {
          ok(true, "getServiced returned the correct value.");
          res();
        } else {
          ok(false, "Wrong value. Expected: " + opened.length +
                    ", got: " + e.data);
          res();
        }
      }
    });

    var content = document.getElementById("content");
    ok(content, "Parent exists.");

    var iframe = document.createElement("iframe");
    iframe.setAttribute('src', "sw_clients/service_worker_controlled.html");
    content.appendChild(iframe);

    return p;
  }

  function removeAndTest() {
    var content = document.getElementById("content");
    ok(content, "Parent exists.");

    content.removeChild(opened.pop());
    content.removeChild(opened.pop());

    return testGetServiced();
  }

  function runTest() {
    start()
      .then(testGetServiced)
      .then(testGetServiced)
      .then(testGetServiced)
      .then(removeAndTest)
      .then(closeAndUnregister)
      .catch(function(e) {
        ok(false, "Some test failed with error " + e);
      }).then(function() {
        SimpleTest.finish();
      });
  }

  SimpleTest.waitForExplicitFinish();
  SpecialPowers.pushPrefEnv({"set": [
    ["dom.serviceWorkers.enabled", true],
    ["dom.serviceWorkers.testing.enabled", true]
  ]}, runTest);
</script>
</pre>
</body>
</html>
